backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=3.7.0
django-filter>=2.4.0
python-dateutil>=2.8.1
ua-parser>=0.18.0
user-agents>=2.2.0
python-ipware>=3.0.0
qrcode>=7.3.1
Pillow>=8.2.0
requests>=2.25.1
beautifulsoup4>=4.9.3
django-anymail>=10.0 